requirements.txt
asgiref==3.8.1
certifi==2024.8.30
channels==4.1.0
charset-normalizer==3.4.0
Django==5.1.2
django-cors-headers==4.5.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
env==0.1.0
idna==3.10
Markdown==3.7
pillow==10.4.0
psycopg2-binary==2.9.10
requests==2.32.3
sqlparse==0.5.1
sslcommerz-lib==1.0
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.2
wrapt==1.11.2